pyproject.toml
[tool.poetry]
name = "pet-math-api"
version = "0.1.0"
description = "FastAPI REST api"
authors = ["AlexeyGrushev <yablokov2004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.104.1"
alembic = "^1.12.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.0.3"
asyncpg = "^0.29.0"
uvicorn = "^0.24.0.post1"
greenlet = "^3.0.1"
openpyxl = "^3.1.2"
email-validator = "^2.1.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "4.0.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
python-multipart = "^0.0.9"
aiofiles = "^23.2.1"
celery = "^5.4.0"
jinja2 = "^3.1.3"
flower = "^2.0.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
